+++ src/main/java/org/urbcomp/startdb/selfstar/utils/Huffman/Code.java
@@ -1,9 +1,9 @@
 package org.urbcomp.startdb.selfstar.utils.Huffman;
 
 public class Code {
-    public final long value;
+    public final int value;
     public final int length;
-    public Code(long value, int length) {
+    public Code(int value, int length) {
         this.value = value;
         this.length = length;
     }
